More bugs fixed

+++ hybrid.sh
@@ -825,8 +825,8 @@ debug_mode_toggle(){
 	echo -n "> "
 	read debug_mode_toggle_opt
 	case $debug_mode_toggle_opt in
-		e|E ) userdebug_cfg && echo "1" > $userdebugcfg && var && echo "Ok!" && sleep 1 && options;;
-		d|D ) userdebug_cfg && echo "0" > $userdebugcfg && var && echo "Ok!" && sleep 1 && options;;
+		e|E ) userdebug_cfg && echo "1" > $userdebugcfg && echo "Ok!" && sleep 1 && $SHELL -c hybrid;;
+		d|D ) userdebug_cfg && echo "0" > $userdebugcfg && echo "Ok!" && sleep 1 && $SHELL -c hybrid;;
 		* ) error_404 && debug_mode_toggle;;
 	esac
 }
@@ -850,8 +850,8 @@ usage_mode_toggle(){
 	echo -n "> "
 	read usage_mode_toggle_opt
 	case $usage_mode_toggle_opt in
-		t|T ) usagetype_cfg && echo "0" > $usagetypecfg && var && echo "Ok!" && sleep 1 && options;;
-		p|P ) usagetype_cfg && echo "1" > $usagetypecfg && var && echo "Ok!" && sleep 1 && options;;
+		t|T ) usagetype_cfg && echo "0" > $usagetypecfg && echo "Ok!" && sleep 1 && $SHELL -c hybrid;;
+		p|P ) usagetype_cfg && echo "1" > $usagetypecfg && echo "Ok!" && sleep 1 && $SHELL -c hybrid;;
 		* ) error_404 && usage_mode_toggle;;
 	esac
 }
@@ -868,19 +868,20 @@ shell_mode_toggle(){
 	fi
 
 	echo "${yellow}Shell Mode:${nc}"
-	echo "if you see random characters in the titles or errors, enable this may help.
+	echo "if you see random characters"
+	echo "in the titles or errors"
+	echo "enable this may help."
 	echo
 	echo "E|Enable"
 	echo "D|Disable"
 	echo
 	echo "${yellow}Currently:${nc} $shfix_status"
 	echo -n "> "
-	echo
 	read shfix_mode_toggle_opt
 	case $shfix_mode_toggle_opt in
-		t|T ) shfix_cfg && echo "0" > $shfixcfg && var && echo "Ok!" && sleep 1 && $SHELL -c hybrid;;
-		p|P ) shfix_cfg && echo "1" > $shfixcfg && var && echo "Ok!" && sleep 1 && $SHELL -c hybrid;;
-		* ) error_404 && usage_mode_toggle;;
+		e|E ) shfix_cfg && echo "1" > $shfixcfg && echo "Ok!" && sleep 1 && $SHELL -c hybrid;;
+		d|D ) shfix_cfg && echo "0" > $shfixcfg && echo "Ok!" && sleep 1 && $SHELL -c hybrid;;
+		* ) error_404 && shell_mode_toggle;;
 	esac
 }
 
@@ -961,8 +962,8 @@ usagetype_first_start(){
 	echo -ne "> "
 	read usagetype_first_start_opt
 	case $usagetype_first_start_opt in
-		t|T ) usagetype_cfg && echo "0" > $usagetypecfg && var && echo "Ok!" && sleep 1 && body;;
-		p|P ) usagetype_cfg && echo "1" > $usagetypecfg && var && echo "Ok!" && sleep 1 && body;;
+		t|T ) usagetype_cfg && echo "0" > $usagetypecfg && echo "Ok!" && sleep 1 && $SHELL -c hybrid;;
+		p|P ) usagetype_cfg && echo "1" > $usagetypecfg && echo "Ok!" && sleep 1 && $SHELL -c hybrid;;
 		* ) error_404 && usagetype_first_start;;
 	esac
 }
@@ -997,26 +998,32 @@ session_behaviour(){
 
 	#check shfix.cfg
 	if [ -e $shfixcfg ]; then
-		elif [ "`grep 0 $shfixcfg`" ]; then
+		if [ "`grep 0 $shfixcfg`" ]; then
 			shfix=0
-		elif [ "`grep 1 $shfixcfg`" ]; then
+		fi
+		if [ "`grep 1 $shfixcfg`" ]; then
 			shfix=1
+		fi
 	fi
 
 	#check usagetype.cfg
 	if [ -e $usagetypecfg ]; then
-		elif [ "`grep 0 $usagetypecfg`" ]; then
+		if [ "`grep 0 $usagetypecfg`" ]; then
 			usagetype=0
-		elif [ "`grep 1 $usagetypecfg`" ]; then
+		fi
+		if [ "`grep 1 $usagetypecfg`" ]; then
 			usagetype=1
+		fi
 	fi
 
 	#check userdebug.cfg
 	if [ -e $userdebugcfg ]; then
-		elif [ "`grep 0 $userdebugcfg`" ]; then
+		if [ "`grep 0 $userdebugcfg`" ]; then
 			userdebug=0
-		elif [ "`grep 1 $userdebugcfg`" ]; then
+		fi
+		if [ "`grep 1 $userdebugcfg`" ]; then
 			userdebug=1
+		fi
 	fi
 
 	#run conditional statements
@@ -1045,11 +1052,13 @@ shfix_session_behaviour(){
 	
 	#run with default Android Shell
 	shfixtmp="/data/local/tmp/shfix.tmp"
-	if [ "`grep 1 $shfixtmp`" ]; then
-		echo "0" > $shfixtmp
-		main_functions
+	if [ -e $shfixtmp ]; then
+		if [ "`grep 1 $shfixtmp`" ]; then
+			echo "0" > $shfixtmp
+			main_functions
+		fi
 	fi
-	
+
 	touch $shfixtmp
 	chmod 755 $shfixtmp
 	echo "0" > $shfixtmp
